requirements.txt
flask==3.0.3
textblob==0.18.0
scikit-learn==1.5.1
nltk==3.8.1
google-generativeai
